Refactoring Added Maven Framework for Dependency management
