+++ src/main/java/src/de/oelrichsgarcia/studipTelegramBot/studip/LoginException.java
@@ -1,4 +1,4 @@
-package studip;
+package src.de.oelrichsgarcia.studipTelegramBot.studip;
 
 /**
  * The type Login exception.
