+++ src/main/java/src/de/oelrichsgarcia/studipTelegramBot/studip/NotLoggedInException.java
@@ -1,4 +1,4 @@
-package studip;
+package src.de.oelrichsgarcia.studipTelegramBot.studip;
 
 public class NotLoggedInException extends AbstractBotException {
