+++ src/main/java/src/de/oelrichsgarcia/studipTelegramBot/studip/api/request/AuthenticatedAPIRequest.java
@@ -1,7 +1,7 @@
-package studip.api.request;
+package src.de.oelrichsgarcia.studipTelegramBot.studip.api.request;
 
 import org.apache.commons.codec.binary.Base64;
-import studip.api.types.Credentials;
+import src.de.oelrichsgarcia.studipTelegramBot.studip.api.types.Credentials;
 
 import javax.net.ssl.HttpsURLConnection;
 import java.io.BufferedReader;
