+++ src/main/java/src/de/oelrichsgarcia/studipTelegramBot/studip/api/types/Uni.java
@@ -1,4 +1,4 @@
-package studip.api.types;
+package src.de.oelrichsgarcia.studipTelegramBot.studip.api.types;
 
 import java.net.URL;
 import java.util.ArrayList;
