+++ src/main/java/src/de/oelrichsgarcia/studipTelegramBot/studip/api/types/User.java
@@ -1,4 +1,4 @@
-package studip.api.types;
+package src.de.oelrichsgarcia.studipTelegramBot.studip.api.types;
 
 import java.util.ArrayList;
